+++ MK4duo/src/core/commands/gcode/config/m569.h
@@ -29,9 +29,20 @@
 #define CODE_M569
 
 /**
- * M569: Set Stepper direction
+ * M569: Stepper driver control
+ *
+ *  X[bool]           - Set X direction
+ *  Y[bool]           - Set Y direction
+ *  Z[bool]           - Set Z direction
+ *  T[tools] E[bool]  - Set Extruder direction
+ *
+ *  P[int]            - Set minimum pulse
+ *  R[long]           - Set maximum rate
+ *
  */
 inline void gcode_M569(void) {
+
+  //#define DEBUG_PULSE_CYCLE
 
   if (commands.get_target_tool(569)) return;
 
@@ -42,13 +53,19 @@ inline void gcode_M569(void) {
     }
   }
 
-  SERIAL_EM("Reporting Stepper Direction");
+  if (parser.seen('P')) stepper.minimum_pulse = parser.value_byte();
+  if (parser.seen('R')) stepper.maximum_rate = parser.value_ulong();
+
+  // Recalculate pulse cycle
+  HAL_calc_pulse_cycle();
+
+  SERIAL_EM("Reporting Stepper control");
   SERIAL_MT(" X dir:", stepper.isStepDir(X_AXIS) ? "true" : "false");
   SERIAL_MT(" Y dir:", stepper.isStepDir(Y_AXIS) ? "true" : "false");
   SERIAL_MV(" Z dir:", stepper.isStepDir(Z_AXIS) ? "true" : "false");
 
   #if DRIVER_EXTRUDERS == 1
-    SERIAL_EMV(" E dir:", stepper.isStepDir(E_AXIS) ? "true" : "false");
+    SERIAL_MV(" E dir:", stepper.isStepDir(E_AXIS) ? "true" : "false");
   #endif
 
   #if DRIVER_EXTRUDERS > 1
@@ -62,5 +79,24 @@ inline void gcode_M569(void) {
     }
   #endif
 
+  SERIAL_MV(" Minimum pulse(us):", stepper.minimum_pulse);
+  SERIAL_MV(" Maximum rate(Hz):", stepper.maximum_rate);
+
   SERIAL_EOL();
+
+  #if ENABLED(DEBUG_PULSE_CYCLE)
+    SERIAL_EMV("HAL_min_pulse_cycle:", HAL_min_pulse_cycle);
+    SERIAL_EMV("HAL_min_pulse_tick:", HAL_min_pulse_tick);
+    SERIAL_EMV("HAL_add_pulse_ticks:", HAL_add_pulse_ticks);
+    SERIAL_EMV("HAL_min_isr_frequency:", HAL_min_isr_frequency);
+    SERIAL_EMV("HAL_frequency_limit[0]:", HAL_frequency_limit[0]);
+    SERIAL_EMV("HAL_frequency_limit[1]:", HAL_frequency_limit[1]);
+    SERIAL_EMV("HAL_frequency_limit[2]:", HAL_frequency_limit[2]);
+    SERIAL_EMV("HAL_frequency_limit[3]:", HAL_frequency_limit[3]);
+    SERIAL_EMV("HAL_frequency_limit[4]:", HAL_frequency_limit[4]);
+    SERIAL_EMV("HAL_frequency_limit[5]:", HAL_frequency_limit[5]);
+    SERIAL_EMV("HAL_frequency_limit[6]:", HAL_frequency_limit[6]);
+    SERIAL_EMV("HAL_frequency_limit[7]:", HAL_frequency_limit[7]);
+  #endif
+
 }